requirements.txt
PyPDF2
PyYAML
streamlit